Create 0x3bEa1Cb0Ea10A35B1995233D6d57B34Fb3142b1B.json
Moiom (MIM) utility token is the currency/money of Africahead, used in conjunction with Africahead Ipparts (AFA) brand token. MIM can be used by anyone for a DApps currency. Part of the project is forming an Ipparts Exchange, where other businesses will also list their brand tokens. MIM, when it is liquid, will be a currency, used to invest in others' brand tokens, on Ipparts Exchange, to, in this way, fund new projects and economic development.